Rename schema.js to definitions.js
