+++ framework/src/modules/chain/api/ws/workers/middlewares/handshake.js
@@ -18,7 +18,7 @@ const url = require('url');
 const _ = require('lodash');
 const failureCodes = require('../../rpc/failure_codes.js');
 const ZSchema = require('../../../../helpers/z_schema');
-const definitions = require('../../../../schema/schema');
+const definitions = require('../../../../schema/definitions');
 const Peer = require('../../../../logic/peer.js');
 
 const validator = new ZSchema();
