+++ framework/src/modules/chain/api/ws/workers/peers_update_rules.js
@@ -18,7 +18,7 @@ const Peer = require('../../../logic/peer');
 const failureCodes = require('../rpc/failure_codes');
 const PeerUpdateError = require('../rpc/failure_codes').PeerUpdateError;
 const ZSchema = require('../../../helpers/z_schema');
-const definitions = require('../../../schema/schema');
+const definitions = require('../../../schema/definitions');
 const connectionsTable = require('./connections_table');
 const SlaveToMasterSender = require('./slave_to_master_sender');
 const Rules = require('./rules');
